Separert ut stans/gjenoppta i egne filer. Service/repo/test.
