+++ database/src/main/kotlin/no/nav/su/se/bakover/database/revurdering/RevurderingPostgresRepo.kt
@@ -7,11 +7,8 @@ import kotliquery.Row
 import no.nav.su.se.bakover.common.objectMapper
 import no.nav.su.se.bakover.common.periode.Periode
 import no.nav.su.se.bakover.common.persistence.SessionFactory
-import no.nav.su.se.bakover.common.persistence.TransactionContext
 import no.nav.su.se.bakover.common.serialize
 import no.nav.su.se.bakover.database.DbMetrics
-import no.nav.su.se.bakover.database.PostgresSessionContext.Companion.withSession
-import no.nav.su.se.bakover.database.PostgresTransactionContext.Companion.withTransaction
 import no.nav.su.se.bakover.database.Session
 import no.nav.su.se.bakover.database.beregning.PersistertBeregning
 import no.nav.su.se.bakover.database.grunnlag.BosituasjongrunnlagPostgresRepo
@@ -95,6 +92,20 @@ internal class RevurderingPostgresRepo(
 ) : RevurderingRepo {
     private val vedtakRepo = VedtakPostgresRepo(dataSource, søknadsbehandlingRepo, this, dbMetrics)
 
+    private val stansAvYtelseRepo = StansAvYtelsePostgresRepo(
+        fradragsgrunnlagPostgresRepo = fradragsgrunnlagPostgresRepo,
+        bosituasjonsgrunnlagPostgresRepo = bosituasjonsgrunnlagPostgresRepo,
+        uføreVilkårsvurderingRepo = uføreVilkårsvurderingRepo,
+        formueVilkårsvurderingRepo = formueVilkårsvurderingRepo,
+    )
+
+    private val gjenopptakAvYtelseRepo = GjenopptakAvYtelsePostgresRepo(
+        fradragsgrunnlagPostgresRepo = fradragsgrunnlagPostgresRepo,
+        bosituasjonsgrunnlagPostgresRepo = bosituasjonsgrunnlagPostgresRepo,
+        uføreVilkårsvurderingRepo = uføreVilkårsvurderingRepo,
+        formueVilkårsvurderingRepo = formueVilkårsvurderingRepo,
+    )
+
     override fun hent(id: UUID): AbstraktRevurdering? {
         return dbMetrics.timeQuery("hentRevurdering") {
             dataSource.withSession { session ->
@@ -127,8 +138,18 @@ internal class RevurderingPostgresRepo(
     override fun lagre(revurdering: AbstraktRevurdering) {
         when (revurdering) {
             is Revurdering -> dataSource.withSession { lagre(revurdering, it) }
-            is GjenopptaYtelseRevurdering -> sessionFactory.withTransactionContext { lagre(revurdering, it) }
-            is StansAvYtelseRevurdering -> sessionFactory.withTransactionContext { lagre(revurdering, it) }
+            is GjenopptaYtelseRevurdering -> sessionFactory.withTransactionContext {
+                gjenopptakAvYtelseRepo.lagre(
+                    revurdering,
+                    it,
+                )
+            }
+            is StansAvYtelseRevurdering -> sessionFactory.withTransactionContext {
+                stansAvYtelseRepo.lagre(
+                    revurdering,
+                    it,
+                )
+            }
         }
     }
 
@@ -140,196 +161,6 @@ internal class RevurderingPostgresRepo(
             is RevurderingTilAttestering -> lagre(revurdering, session)
             is IverksattRevurdering -> lagre(revurdering, session)
             is UnderkjentRevurdering -> lagre(revurdering, session)
-        }
-    }
-
-    internal fun lagre(revurdering: StansAvYtelseRevurdering, sessionContext: TransactionContext) {
-        sessionContext.withTransaction { tx ->
-            when (revurdering) {
-                is StansAvYtelseRevurdering.SimulertStansAvYtelse -> {
-                    """
-                    insert into revurdering (
-                        id,
-                        opprettet,
-                        periode,
-                        simulering,
-                        saksbehandler,
-                        revurderingsType,
-                        vedtakSomRevurderesId,
-                        årsak,
-                        begrunnelse,
-                        attestering,
-                        skalFøreTilBrevutsending
-                    ) values (
-                        :id,
-                        :opprettet,
-                        to_json(:periode::json),
-                        to_json(:simulering::json),
-                        :saksbehandler,
-                        :revurderingsType,
-                        :vedtakSomRevurderesId,
-                        :arsak,
-                        :begrunnelse,
-                        to_json(:attestering::json),
-                        :skalFoereTilBrevutsending
-                    ) on conflict(id) do update set
-                        periode=to_json(:periode::json),
-                        simulering=to_json(:simulering::json),
-                        saksbehandler=:saksbehandler,
-                        revurderingsType=:revurderingsType,
-                        vedtakSomRevurderesId=:vedtakSomRevurderesId,
-                        årsak=:arsak,
-                        begrunnelse=:begrunnelse
-                    """.trimIndent()
-                        .insert(
-                            mapOf(
-                                "id" to revurdering.id,
-                                "opprettet" to revurdering.opprettet,
-                                "periode" to objectMapper.writeValueAsString(revurdering.periode),
-                                "simulering" to objectMapper.writeValueAsString(revurdering.simulering),
-                                "saksbehandler" to revurdering.saksbehandler,
-                                "revurderingsType" to RevurderingsType.SIMULERT_STANS,
-                                "vedtakSomRevurderesId" to revurdering.tilRevurdering.id,
-                                "arsak" to revurdering.revurderingsårsak.årsak.toString(),
-                                "begrunnelse" to revurdering.revurderingsårsak.begrunnelse.toString(),
-                                "attestering" to Attesteringshistorikk.empty().hentAttesteringer().serialize(),
-                                "skalFoereTilBrevutsending" to false,
-                            ),
-                            tx,
-                        )
-                    fradragsgrunnlagPostgresRepo.lagreFradragsgrunnlag(
-                        behandlingId = revurdering.id,
-                        fradragsgrunnlag = revurdering.grunnlagsdata.fradragsgrunnlag,
-                        tx = tx,
-                    )
-                    bosituasjonsgrunnlagPostgresRepo.lagreBosituasjongrunnlag(
-                        behandlingId = revurdering.id,
-                        grunnlag = revurdering.grunnlagsdata.bosituasjon,
-                        tx = tx,
-                    )
-                    uføreVilkårsvurderingRepo.lagre(
-                        behandlingId = revurdering.id,
-                        vilkår = revurdering.vilkårsvurderinger.uføre,
-                        tx = tx,
-                    )
-                    formueVilkårsvurderingRepo.lagre(
-                        behandlingId = revurdering.id,
-                        vilkår = revurdering.vilkårsvurderinger.formue,
-                        tx = tx,
-                    )
-                }
-                is StansAvYtelseRevurdering.IverksattStansAvYtelse -> {
-                    """
-                    update revurdering set 
-                        attestering = to_json(:attestering::json),
-                        revurderingsType = :revurderingsType 
-                    where id = :id
-                    """.trimIndent()
-                        .oppdatering(
-                            mapOf(
-                                "attestering" to revurdering.attesteringer.hentAttesteringer().serialize(),
-                                "revurderingsType" to RevurderingsType.IVERKSATT_STANS,
-                                "id" to revurdering.id,
-                            ),
-                            tx,
-                        )
-                }
-            }
-        }
-    }
-
-    internal fun lagre(revurdering: GjenopptaYtelseRevurdering, sessionContext: TransactionContext) {
-        sessionContext.withTransaction { tx ->
-            when (revurdering) {
-                is GjenopptaYtelseRevurdering.SimulertGjenopptakAvYtelse -> {
-                    """
-                    insert into revurdering (
-                        id,
-                        opprettet,
-                        periode,
-                        simulering,
-                        saksbehandler,
-                        revurderingsType,
-                        vedtakSomRevurderesId,
-                        årsak,
-                        begrunnelse,
-                        attestering,
-                        skalFøreTilBrevutsending
-                    ) values (
-                        :id,
-                        :opprettet,
-                        to_json(:periode::json),
-                        to_json(:simulering::json),
-                        :saksbehandler,
-                        :revurderingsType,
-                        :vedtakSomRevurderesId,
-                        :arsak,
-                        :begrunnelse,
-                        to_json(:attestering::json),
-                        :skalFoereTilBrevutsending
-                    ) on conflict(id) do update set
-                        periode=to_json(:periode::json),
-                        simulering=to_json(:simulering::json),
-                        saksbehandler=:saksbehandler,
-                        revurderingsType=:revurderingsType,
-                        vedtakSomRevurderesId=:vedtakSomRevurderesId,
-                        årsak=:arsak,
-                        begrunnelse=:begrunnelse
-                    """.trimIndent()
-                        .insert(
-                            mapOf(
-                                "id" to revurdering.id,
-                                "opprettet" to revurdering.opprettet,
-                                "periode" to objectMapper.writeValueAsString(revurdering.periode),
-                                "simulering" to objectMapper.writeValueAsString(revurdering.simulering),
-                                "saksbehandler" to revurdering.saksbehandler,
-                                "revurderingsType" to RevurderingsType.SIMULERT_GJENOPPTAK,
-                                "vedtakSomRevurderesId" to revurdering.tilRevurdering.id,
-                                "arsak" to revurdering.revurderingsårsak.årsak.toString(),
-                                "begrunnelse" to revurdering.revurderingsårsak.begrunnelse.toString(),
-                                "attestering" to Attesteringshistorikk.empty().hentAttesteringer().serialize(),
-                                "skalFoereTilBrevutsending" to false,
-                            ),
-                            tx,
-                        )
-                    fradragsgrunnlagPostgresRepo.lagreFradragsgrunnlag(
-                        behandlingId = revurdering.id,
-                        fradragsgrunnlag = revurdering.grunnlagsdata.fradragsgrunnlag,
-                        tx = tx,
-                    )
-                    bosituasjonsgrunnlagPostgresRepo.lagreBosituasjongrunnlag(
-                        behandlingId = revurdering.id,
-                        grunnlag = revurdering.grunnlagsdata.bosituasjon,
-                        tx = tx,
-                    )
-                    uføreVilkårsvurderingRepo.lagre(
-                        behandlingId = revurdering.id,
-                        vilkår = revurdering.vilkårsvurderinger.uføre,
-                        tx = tx,
-                    )
-                    formueVilkårsvurderingRepo.lagre(
-                        behandlingId = revurdering.id,
-                        vilkår = revurdering.vilkårsvurderinger.formue,
-                        tx = tx,
-                    )
-                }
-                is GjenopptaYtelseRevurdering.IverksattGjenopptakAvYtelse -> {
-                    """
-                    update revurdering set 
-                        attestering = to_json(:attestering::json),
-                        revurderingsType = :revurderingsType 
-                    where id = :id
-                    """.trimIndent()
-                        .oppdatering(
-                            mapOf(
-                                "attestering" to revurdering.attesteringer.hentAttesteringer().serialize(),
-                                "revurderingsType" to RevurderingsType.IVERKSATT_GJENOPPTAK,
-                                "id" to revurdering.id,
-                            ),
-                            tx,
-                        )
-                }
-            }
         }
     }
