+++ service/src/main/kotlin/no/nav/su/se/bakover/service/revurdering/StansAvYtelseService.kt
@@ -1,0 +1,122 @@
+package no.nav.su.se.bakover.service.revurdering
+
+import arrow.core.Either
+import arrow.core.getOrHandle
+import arrow.core.left
+import arrow.core.right
+import no.nav.su.se.bakover.common.Tidspunkt
+import no.nav.su.se.bakover.database.revurdering.RevurderingRepo
+import no.nav.su.se.bakover.database.vedtak.VedtakRepo
+import no.nav.su.se.bakover.domain.NavIdentBruker
+import no.nav.su.se.bakover.domain.behandling.Attestering
+import no.nav.su.se.bakover.domain.revurdering.StansAvYtelseRevurdering
+import no.nav.su.se.bakover.domain.vedtak.GjeldendeVedtaksdata
+import no.nav.su.se.bakover.domain.vedtak.Vedtak
+import no.nav.su.se.bakover.service.utbetaling.UtbetalingService
+import no.nav.su.se.bakover.service.vedtak.VedtakService
+import java.time.Clock
+import java.util.UUID
+
+internal class StansAvYtelseService(
+    private val utbetalingService: UtbetalingService,
+    private val revurderingRepo: RevurderingRepo,
+    private val clock: Clock,
+    private val vedtakRepo: VedtakRepo,
+    private val vedtakService: VedtakService,
+) {
+    fun stansAvYtelse(
+        request: StansYtelseRequest,
+    ): Either<KunneIkkeStanseYtelse, StansAvYtelseRevurdering.SimulertStansAvYtelse> {
+
+        val gjeldendeVedtaksdata: GjeldendeVedtaksdata = vedtakService.kopierGjeldendeVedtaksdata(
+            sakId = request.sakId,
+            fraOgMed = request.fraOgMed,
+        ).getOrHandle {
+            return KunneIkkeStanseYtelse.KunneIkkeOppretteRevurdering.left()
+        }.also {
+            if (!it.tidslinjeForVedtakErSammenhengende()) return KunneIkkeStanseYtelse.KunneIkkeOppretteRevurdering.left()
+        }
+
+        val simulering = utbetalingService.simulerStans(
+            sakId = request.sakId,
+            saksbehandler = request.saksbehandler,
+            stansDato = request.fraOgMed,
+        ).getOrHandle {
+            return KunneIkkeStanseYtelse.SimuleringAvStansFeilet(it).left()
+        }
+
+        val simulertRevurdering = when (request) {
+            is StansYtelseRequest.Oppdater -> {
+                val eksisterende = revurderingRepo.hent(request.revurderingId)
+                    ?: return KunneIkkeStanseYtelse.FantIkkeRevurdering.left()
+                when (eksisterende) {
+                    is StansAvYtelseRevurdering.SimulertStansAvYtelse -> {
+                        eksisterende.copy(
+                            periode = gjeldendeVedtaksdata.periode,
+                            grunnlagsdata = gjeldendeVedtaksdata.grunnlagsdata,
+                            vilkårsvurderinger = gjeldendeVedtaksdata.vilkårsvurderinger,
+                            tilRevurdering = gjeldendeVedtaksdata.gjeldendeVedtakPåDato(request.fraOgMed)!!,
+                            saksbehandler = request.saksbehandler,
+                            simulering = simulering.simulering,
+                            revurderingsårsak = request.revurderingsårsak,
+                        )
+                    }
+                    else -> return KunneIkkeStanseYtelse.UgyldigTypeForOppdatering(eksisterende::class).left()
+                }
+            }
+            is StansYtelseRequest.Opprett -> {
+                StansAvYtelseRevurdering.SimulertStansAvYtelse(
+                    id = UUID.randomUUID(),
+                    opprettet = Tidspunkt.now(),
+                    periode = gjeldendeVedtaksdata.periode,
+                    grunnlagsdata = gjeldendeVedtaksdata.grunnlagsdata,
+                    vilkårsvurderinger = gjeldendeVedtaksdata.vilkårsvurderinger,
+                    tilRevurdering = gjeldendeVedtaksdata.gjeldendeVedtakPåDato(request.fraOgMed)!!,
+                    saksbehandler = request.saksbehandler,
+                    simulering = simulering.simulering,
+                    revurderingsårsak = request.revurderingsårsak,
+                )
+            }
+        }
+
+        revurderingRepo.lagre(simulertRevurdering)
+
+        return simulertRevurdering.right()
+    }
+
+    fun iverksettStansAvYtelse(
+        revurderingId: UUID,
+        attestant: NavIdentBruker.Attestant,
+    ): Either<KunneIkkeIverksetteStansYtelse, StansAvYtelseRevurdering.IverksattStansAvYtelse> {
+        val revurdering = revurderingRepo.hent(revurderingId)
+            ?: return KunneIkkeIverksetteStansYtelse.FantIkkeRevurdering.left()
+
+        return when (revurdering) {
+            is StansAvYtelseRevurdering.SimulertStansAvYtelse -> {
+                val iverksattRevurdering = revurdering.iverksett(
+                    Attestering.Iverksatt(
+                        attestant,
+                        Tidspunkt.now(clock),
+                    ),
+                )
+
+                val stansUtbetaling = utbetalingService.stansUtbetalinger(
+                    sakId = iverksattRevurdering.sakId,
+                    attestant = iverksattRevurdering.attesteringer.hentSisteAttestering().attestant,
+                    simulering = revurdering.simulering,
+                    stansDato = iverksattRevurdering.periode.fraOgMed,
+                ).getOrHandle { return KunneIkkeIverksetteStansYtelse.KunneIkkeUtbetale(it).left() }
+
+                val vedtak = Vedtak.from(iverksattRevurdering, stansUtbetaling.id, clock)
+
+                revurderingRepo.lagre(iverksattRevurdering)
+                vedtakRepo.lagre(vedtak)
+
+                return iverksattRevurdering.right()
+            }
+            else -> KunneIkkeIverksetteStansYtelse.UgyldigTilstand(
+                faktiskTilstand = revurdering::class,
+            ).left()
+        }
+    }
+}
